ingest-client/requirements.txt
watchdog>=4.0
requests>=2.31
python-dotenv>=1.0
